Set default MIDI instruments for each voice
Because the Sinfonias each have three voices, it is nice to be able to
change the MIDI instrument for each voice to help learn each part, to
experiment with different arrangements, etc.

Set the MIDI instruments to "acoustic grand" (in global variables) as
appropriate production default values, and use these variables in each
\midi block. We can use three Staff contexts to facilitate this. Then,
instruments can be changed easily by modifying the variables and
regenerating the MIDI files.