Sinfonia 4: proofreading
Measure 8, beat 4, upper part: The F-natural should be tied over to the
F-natural in the next measure.

Measure 12, beat 4, lower part: The second and fourth sixteenth notes
should be G-sharps.

Measure 14, beat 3, middle part: The F-sharp in this beat should be a
D-sharp (lower chromatic neighbor tone instead of an upper neighbor).

Measure 16, beat 3, upper part: The A-natural in this beat should
instead be a C-natural.

Measure 16, lower part: Revise the entire measure: correct note
durations of the first four notes, and add missing sixteenth notes in
the last two beats.

Measure 17, beat 2, lower part: The first eighth note in this beat
should be E instead of F.

Measure 18, beat 2, upper part: The fourth sixteenth note should be a
D instead of F.