Sinfonia 8: proofreading
Measure 7, beat 1, middle part: The first eighth note (with \prall)
should be F-natural, and the second sixteenth should be D-natural.

Measure 19, beat 3, lower part: Listed sources disagree whether the
third sixteenth note of this beat should be A or D. The manuscript has
an A, so change the note to A.

Measure 20, beat 2, lower part: The fourth sixteenth note of this beat
should be A instead of B-flat.

Measure 21, beat 3, upper part: The quarter note F in this beat is
correctly tied from the previous F, but it should also be tied to the
sixteenth-note F in the next beat.