Sinfonia 9: proofreading
Measure 2, beat 3, middle part: The second sixteenth note should be
D-flat instead of D-natural.

Measure 3, beats 2 and 3, lower part: The A-naturals in these two beats
should be A-flats.

Last measure, middle part: The two listed sources for this edition
disagree, but Bach's manuscript (and various urtext editions I checked)
write A-natural instead of A-flat, making it a Picardy third. I opted to
follow the manuscript in this case.