Sinfonia 10: proofreading
Measure 24, beat 2, lower part: The two listed sources disagree about
whether the sixteenth-note G in this beat should be G-sharp or
G-natural. Bach's manuscript does not have an accidental for the G,
implying a G-natural, and other urtext editions I checked add the
natural sign. In interest of trusting the manuscript, I changed the note
to G-natural.

Measure 28, beat 3, lower part: The second sixteenth note of this beat
should be D instead of E.