Sinfonia 14: proofreading
Measure 3, beat 2, lower part: The last sixteenth note of this beat
should be G instead of F.

Measure 7, beat 1, lower part: The two listed sources disagree about
whether the last sixteenth of this beat should be F or D. Consulting the
manuscript and a couple urtext editions, those editions agree that the
note is D.

Measure 7, beat 2, lower part: Fix the MIDI version of the trill; the
principal note is E-flat instead of E-natural.

Measure 15, beat 4, lower part: The two listed sources disagree about
the note sequence for this beat. Source 1 has { C, D, C, B }, which this
edition currently follows. Source 2 has { B-flat, D, C, B-flat }, which
is the sequence used in a few urtext editions I've consulted.  However,
Bach's manuscript as { B, D, C, B }. I have yet to find commentary to
explain why the urtext editions diverge from the manuscript, but I
suspect the scholarly consensus is that Bach forgot to add the flat
accidental to first B sixteenth note, which would yield the sequence {
B-flat, D, C, B-flat } that matches the urtext editions. The manuscript
omits a flat accidental on the B in middle part as well, but urtext
editions, and Source 1 and 2, all add the flat. Perhaps there are other
contemporaneous copys of the manuscript that add the B-flats, but I have
not checked for that yet. For now, change the notes to match Source 2
and the urtext editions.