Sinfonia 13: proofreading (round 2)
Measures 18 and 19, lower part: Source 1 has two repeated F-naturals in
measure 18, and two repeated E-naturals in measure 19, but in the
manuscript, these notes span their entire measure (as dotted-quarter
notes). In my opinion, the editor for the Source 1 edition rewrote these
two measures to facilitate understanding of how to perform the passage,
but in turn, Bach's intention was lost. Source 2 is faithful to the
manuscript, having one dotted-quarter note in both measures, 18 and 19.
Make this change to reflect Source 2 and the manuscript.

Measures 18 and 19, middle part: Use the Source 2 layout for staff
change and clef change in the middle voice. With this layout, the
parenthesized note is no longer needed, and the notation is easier to
read with the changes made to the lower part (see previous paragraph).